feature-store/python/darwin_fs/pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "darwin_fs"
version = "2.0.3"
dependencies = ["requests~=2.32.3", "pyyaml~=6.0.2", "dataclasses-json~=0.5.7"]

[project.optional-dependencies]
all = ["pyspark>=3.3.1,<=3.5.4", "aiohttp>=3.7"]
spark = ["pyspark>=3.3.1,<=3.5.4"]
async = ["aiohttp>=3.7"]
test = ["pyspark>=3.3.1,<=3.5.4", "aiohttp>=3.7", "asyncio", "wiremock==2.6.1", "pytest", "pytest-asyncio", "testcontainers==4.9.1", "kafka-python", "boto3"]